Update license to LGPL 3.
We should use the same license as Triber. Their GitHub says they use LGPL
and their latest commit on the LICENSE file as well. As does the project
leader. Sadly, the actual license text is GPL, that was used here before.
This resolves the issue by using the actual LGPL text.

All code currently in the repository is written by Jos (signer of this
commit).