Add converters, refactor code.
Also needed to add the Unsafe class, because Jack sometimes
generates opcodes that uses that class. This was recommended
as a workaround. Google will not fix it as Jack is EOL.
However, the recommended Android Studio Preview did not
work.

The Bunq converters is adjusted code from the Square Gson
adapter. Signing and verifying will be added. Sadly there
was no way to do middleware in this system.